Add command to convert poetry.lock to JSON

+++ src/ballad/tests.py
@@ -1,8 +1,13 @@
-import tomlkit
 import json
+import os
+
+import tomlkit
 
 
-def lock_to_json(lock: str="poetry.lock", output: str="out.json"):
+def convert_lock(lock: str = "poetry.lock", output: str = "poetry.json"):
+  if not os.path.isfile(lock):
+    raise FileNotFoundError(f"Cannot find {lock}")
+
   with open(lock, "rt") as f_lock:
     with open(output, "wt") as f_out:
       json.dump(tomlkit.parse(f_lock.read()), f_out, indent=2)
